+++ Watchdogd/Include/Module_dls.h
@@ -33,10 +33,10 @@
  extern int M( int num );
  extern int TR( int num );
  extern int TRCount( int num );
- extern int EA_ech_inf( double val, int num );
- extern int EA_ech_sup( double val, int num );
- extern int EA_ech_inf_egal( double val, int num );
- extern int EA_ech_sup_egal( double val, int num );
+ extern int EA_ech_inf( float val, int num );
+ extern int EA_ech_sup( float val, int num );
+ extern int EA_ech_inf_egal( float val, int num );
+ extern int EA_ech_sup_egal( float val, int num );
  extern float CI( int num );
  extern void SI( int num, int etat, int rouge, int vert, int bleu, int cligno );
  extern void SB( int num, int etat );
